* Attempting to fix sidenav
