+++ frontend/src/app/app.component.html
@@ -5,7 +5,52 @@
     </mat-toolbar>
 
     <mat-sidenav-container class="app-sidenav-container" [style.marginTop.px]="mobileQuery.matches ? 56 : 0">
-        <app-sidenav #snav></app-sidenav>
+        <mat-sidenav #snav [mode]="mobileQuery.matches ? 'over' : 'side'"
+                     [opened]="isSidebarOpenOnPageLoad()"
+                     [fixedInViewport]="mobileQuery.matches" fixedTopGap="56">
+
+            <!--Nested nav based on: https://stackoverflow.com/questions/44374200/angular-2-material-nested-md-list-item-->
+            <mat-nav-list>
+                <div *ngFor="let navItem of navItems">
+
+                    <a *ngIf="!navItem.subItems" href="{{ navItem.link }}">
+                        <mat-list-item class="nav-item"
+                                       [class.selected-nav-item]="navItem.isCurrentlySelected(location.path())">
+                            {{navItem.display}}
+                        </mat-list-item>
+                    </a>
+
+                    <mat-list-item *ngIf="navItem.subItems"
+                                   class="nav-item"
+                                   (click)="navItem.showSubItems = !navItem.showSubItems"
+                                   [class.selected-nav-item]="navItem.isCurrentlySelected(location.path())">
+                        {{navItem.display}}
+                        <i *ngIf="!navItem.showSubItems" class="material-icons text-right">keyboard_arrow_down</i>
+                        <i *ngIf="navItem.showSubItems" class="material-icons text-right">keyboard_arrow_up</i>
+                    </mat-list-item>
+
+                    <!-- Nested navigation elements -->
+                    <mat-nav-list *ngIf="navItem.subItems && navItem.showSubItems"
+                                  style="margin-left:30px;">
+                        <div *ngFor="let subItem of navItem.subItems">
+                            <a href="{{ subItem.link }}">
+                                <mat-list-item
+                                        [class.selected-nav-item]="subItem.isCurrentlySelected(location.path())">
+                                    {{ subItem.display }}
+                                </mat-list-item>
+                            </a>
+                        </div>
+                    </mat-nav-list>
+                </div>
+
+                <div>
+                    <mat-list-item *ngIf="!isInProduction()" class="nav-item developer-nav-item"
+                                   (click)="switchConfigFile()">
+                        Toggle Frontend Config
+                    </mat-list-item>
+                </div>
+            </mat-nav-list>
+        </mat-sidenav>
         <router-outlet></router-outlet>
     </mat-sidenav-container>
 </div>
