--- frontend/src/app/sidenav/sidenav.component.css
@@ -1,52 +1,0 @@
-.app-container {
-    display: flex;
-    flex-direction: column;
-    position: absolute;
-    top: 0;
-    bottom: 0;
-    left: 0;
-    right: 0;
-}
-
-.app-is-mobile .app-toolbar {
-    position: fixed;
-    /* Make sure the toolbar will stay on top of the content as it scrolls past. */
-    z-index: 2;
-}
-
-h1.app-name {
-    margin-left: 8px;
-}
-
-.app-sidenav-container {
-    /* When the sidenav is not fixed, stretch the sidenav container to fill the available space. This
-       causes `<mat-sidenav-content>` to act as our scrolling element for desktop layouts. */
-    flex: 1;
-}
-
-.app-is-mobile .app-sidenav-container {
-    /* When the sidenav is fixed, don't constrain the height of the sidenav container. This allows the
-       `<body>` to be our scrolling element for mobile layouts. */
-    flex: 1 0 auto;
-}
-
-.nav-item {
-    padding-left: 5px;
-    padding-right: 125px;
-}
-
-.selected-nav-item {
-    background-color: #dbdbdb;
-}
-
-.developer-nav-item {
-    background-color: #ffb401;
-}
-
-.developer-nav-item:hover {
-    background-color: #bd8401;
-}
-
-.mat-list-item {
-    user-select: none;
-}
